[FIX] #784 Erreur modification article

+++ articles/controllers/ArticlesDisplayPendingArticlesController.class.php
@@ -235,7 +235,7 @@ class ArticlesDisplayPendingArticlesController extends ModuleController
 		
 		$current_page = AppContext::get_request()->get_getint('page', 1);
 		
-		$pagination = new ModulePagination($current_page, $number_articles, ArticlesConfig::load()->get_number_articles_per_page());
+		$pagination = new ModulePagination($current_page, $number_articles, (int)ArticlesConfig::load()->get_number_articles_per_page());
 		$pagination->set_url(ArticlesUrlBuilder::display_pending_articles($field, $mode, '/%d'));
 		
 		if ($pagination->current_page_is_empty() && $current_page > 1)
